# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Carbon Footprint Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /offsets/purchases:
    get:
      operationId: listOffsetPurchases
      tags: [offsets]
      summary: Get offset purchase history and summary
      responses:
        "200":
          description: Offset summary with purchases
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OffsetSummary"
    post:
      operationId: purchaseOffset
      tags: [offsets]
      summary: Purchase a carbon offset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OffsetPurchaseInput"
      responses:
        "201":
          description: Purchase confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OffsetPurchase"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    OffsetPurchase:
      type: object
      required: [id, projectId, projectName, amountKg, amountUsd, purchasedAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        projectName:
          type: string
        amountKg:
          type: number
          description: "CO2 offset in kg"
        amountUsd:
          type: number
        purchasedAt:
          type: string
          format: date-time
    OffsetPurchaseInput:
      type: object
      required: [projectId, amountKg]
      properties:
        projectId:
          type: integer
        amountKg:
          type: number
    OffsetSummary:
      type: object
      required: [totalOffsetKg, totalSpentUsd, netEmissionsKg, totalEmissionsKg, purchases]
      properties:
        totalOffsetKg:
          type: number
        totalSpentUsd:
          type: number
        netEmissionsKg:
          type: number
        totalEmissionsKg:
          type: number
        purchases:
          type: array
          items:
            $ref: "#/components/schemas/OffsetPurchase"