# lib/api-zod/src/generated/types/offsetSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carbon Footprint Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { OffsetPurchase } from './offsetPurchase';

export interface OffsetSummary {
  totalOffsetKg: number;
  totalSpentUsd: number;
  netEmissionsKg: number;
  totalEmissionsKg: number;
  purchases: OffsetPurchase[];
}